+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -1,24 +1,26 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-public class LottoWinnerBonusNumberTest {
+public class LottoNumberTest {
+
     @Test
-    @DisplayName("보너스 숫자를 생성한다.")
+    @DisplayName("숫자를 생성한다.")
     public void createNumberTest() {
-        LottoWinnerBonusNumber number = new LottoWinnerBonusNumber(5);
-        assertThat(number).isEqualTo(new LottoWinnerBonusNumber(5));
+        LottoNumber number = new LottoNumber(5);
+        assertThat(number).isEqualTo(new LottoNumber(5));
     }
 
     @Test
-    @DisplayName("보너스 로또 숫자는 1~45 사이의 숫자여야한다.")
+    @DisplayName("로또 숫자는 1~45 사이의 숫자여야한다.")
     public void validateNumberTest() {
         assertThatThrownBy(() -> {
-            new LottoWinnerBonusNumber(46);
+            new LottoNumber(46);
         }).isInstanceOf(IllegalArgumentException.class);
     }
 }
