+++ src/test/java/lotto/domain/LottoTicketsTest.java
@@ -1,5 +1,8 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoTicket;
+import lotto.domain.LottoTickets;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
