+++ src/test/java/lotto/domain/LottoWinnerTest.java
@@ -1,5 +1,9 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoWinner;
+import lotto.domain.LottoWinnerBonusNumber;
+import lotto.domain.LottoWinnerTicket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
